fix: save시 id값 받도록 수정

+++ src/main/java/nextstep/courses/domain/session/SessionRepository.java
@@ -1,7 +1,7 @@
 package nextstep.courses.domain.session;
 
 public interface SessionRepository {
-    int save(Session session);
+    long save(Session session);
 
     Session findById(Long id);
 }
